Clean up and adding unit tests, also code review feedback
Delta analysis

HttpResponse
has one method which is a callback, that takes a string body.
It might be better if this took an HttpResponseBody so that you could
hide/abstract the bytes from the server. This would also enable a
binary protocol in the future.
This is not a high priority.

HttpServer
This should have the ability for other types of interception pre/post
for web socket and HTTP requests.

WebsSocketSender
Only supports strings. If we ever go to a binary protocol. This is
limited.
Change send to sendText. Add a binary sentBinary that takes a byte[].
Use a default method to call the sendText from sendBinary so that only
sendText needs to be implemented.

JsonMapper
Add support for Jackson.

Sender
Only supports text see WebsSocketSender delta.

BasicQueue
This class does too much
It needs knowledge of the type of queue.
There should be several queue implementations.
We need another that does metrics.
The queue also spawns a thread and does its own scheduling.
We really need the scheduling done by a third partys so we can add
different types of schedulers.

QueueScheduler
Add this concept and take it out of BasicQueue

ReceiveQueueManager
This will become QueueScheduler

ServiceImpl
There is no way to configure with different method handler.
There is no way to pass AOP interception.
There is no callback for blocked call to be put into another queue or
holding area for data faulting.
add this to doHandleMethodCall and beforeMethodProcessing
ServiceImpl can be run async, i.e., it manages its queue or it can run
sync, i.e., the service bundle manages its queue. This is probably ok.

BoonServiceProxyFactory and BoonJSONServiceFactory seem to do the exact
same thing

ServiceStore
Availability DC 2 DC,
Throughput

ServiceServerImpl
We really need a way to capture pre/post web socket and pre/post request

Testing deltas
BoonJSONServiceFactory has no test and no coverage