QBit : Increased ServiceServerImpl code coverage yesterday.
Increased ServiceServerImpl yesterday.
Changed ReceiveQueueListener to use default methods.

Note that ServiceServerImpl only handles idle callback for checking
timeouts. This may not be good enough if under heavily load use case.
We may need to do a periodic check as well for timeouts. Not only when
we are not busy.