Bag for request specific data and its support
In order to support aspect related operations such as authentication,
session management etc, we need to carry a bag of arbitrary objects on
the request.

When integrated with Vertx, for instance, developers can assign their
own authentication handler that would set some sort of auth context for
the request, which we will pick up in qbit.